refactor: refactored helper functions
